feat: timeout handling for per user permissions
Users in timeout will have all of their per user permissions in gaming,
hobbies and learning languages removed.
Attempts to toggle the reactions twice will be denied.
When the timeout is lifted, the reactions in the three list channels are
scanned for reactions of that user and permissions are given for each
active reaction.