feat: Mechanism to handle ghosting buddies
This commit aims to provide a way for users who were matched in the
buddy project to report that they have not heard back from their buddies
and would like to get matched again.
This is achieved by marking users as ghosted in the database and
running a script periodically to check for users who were reported as
ghosted more than a week ago. Those users and their buddies are
unmatched and the user who reported being ghosted is matched with a new
user.
To provide a way to prevent people just yeeting out their old buddy in
favour of a new one, a message is sent to the buddy asking them to
confirm that they are still using Discord and are willing to message
their buddy within a week by clicking a reaction below the message. When
they click it, their buddy (the user who originally reported being
ghosted) is messaged, letting them know about it and the database entry
is changed to reflect that the user should not be ghosted anymore.