feat: Run requests manually,cleaner error handling
Previous version used discord.js for role assigning because I thought it
would bulk fetch users but there is no endpoint for that so I assume it
runs single requests for each user. We don't need that so the requests
to add the roles are now manual put requests.

The error handling has been cleaned up to shut down the bot on every
possible error instead of select cases.