tickets now use ids instead of the name and tag
The command "/shoutout-perms" now does not require an argument of the
user in order to function. It uses the ID contained in the channel's
name to figure out which user to toggle the shoutout permissions for.

Keep in mind that the change to userIds also applies to other types of
tickets!